Expose dataStoreIdentifier and webPushPartition getters on WKWebsiteDataStore https://bugs.webkit.org/show_bug.cgi?id=252067 rdar://problem/105287995
Reviewed by Sihui Liu.

When showing a notification, we need to provide the embedder with a way to know the data store
identifier and push partition that the notification is associated with. This patch facilitates this
by adding those properties to WKWebsiteDataStore. You could do this before, but it would require
accessing the configuration property on WKWebsiteDataStore, which does a large copy operation.

Additionally, this provides a way to go from a the C API WKWebsiteDataStoreRef to Obj-C API
WKWebsiteDataStore. This is because the callback for showing a notification uses the C API.

Contains miscellaneous unified build fixes.

* Source/WebKit/PlatformMac.cmake:
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/UIProcess/API/C/mac/WKWebsiteDataStoreRefPrivateMac.h: Copied from Source/WebKit/UIProcess/API/Cocoa/_WKInspectorInternal.h.
* Source/WebKit/UIProcess/API/C/mac/WKWebsiteDataStoreRefPrivateMac.mm: Copied from Source/WebKit/UIProcess/API/Cocoa/_WKInspectorInternal.h.
(WKWebsiteDataStoreGetDataStore):
* Source/WebKit/UIProcess/API/Cocoa/APIAttachmentCocoa.mm:
* Source/WebKit/UIProcess/API/Cocoa/WKWebsiteDataStore.mm:
(-[WKWebsiteDataStore _identifier]):
(-[WKWebsiteDataStore _webPushPartition]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebsiteDataStorePrivate.h:
* Source/WebKit/UIProcess/API/Cocoa/_WKInspector.mm:
* Source/WebKit/UIProcess/API/Cocoa/_WKInspectorInternal.h:
* Source/WebKit/UIProcess/API/ios/WKWebViewTestingIOS.mm:
* Source/WebKit/UIProcess/Cocoa/WebProcessProxyCocoa.mm:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebsiteDatastore.mm:
(TestWebKitAPI::createWebsiteDataStoreAndPrepare):
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/260379@main