css/css-transitions/changing-while-transition-004.html is a unique failure https://bugs.webkit.org/show_bug.cgi?id=235132 rdar://87785199
Reviewed by Dean Jackson.

We used to assume that not finding an Animation object matching the CSS property being considered for transitions
meant that that property was not found in the transition-property value list. However, the initial value for
transition-property is `all` and means that any CSS property matches. So we must distinguish between
that and having a matching Animation object to correctly implement the "starting a transition" procedure.

* LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/changing-while-transition-004-expected.txt:
* Source/WebCore/style/Styleable.cpp:
(WebCore::compileTransitionPropertiesInStyle):
(WebCore::updateCSSTransitionsForStyleableAndProperty):
(WebCore::transitionCombinedDuration): Deleted.

Canonical link: https://commits.webkit.org/260398@main