Compute the correct overflow-x and overflow-y values for table elements
Compute the correct overflow-x and overflow-y values for table elements
https://bugs.webkit.org/show_bug.cgi?id=252422

Reviewed by NOBODY (OOPS!).

This patch is to align WebKit with Chromium / Blink and Firefox / Gecko.

Merge - https://src.chromium.org/viewvc/blink?view=revision&revision=199640

This patch is to align WebKit to respect web-specification[1] & [2], which tells us that
display:table must treat anything other than hidden as visible and if overflow-x and
overflow-y are different, how they must be resolved. This latter part is already taken care of,
we're just allowing it to be applied to tables and table parts now.

[1] https://drafts.csswg.org/css2/#overflow
[2] https://drafts.csswg.org/css-overflow-3/#propdef-overflow-x

* Source/WebCore/style/StyleAdjuster.cpp:
(Adjuster::adjust): Update handling of "overflow-x" and "overflow-y" align with spec and
interop with other engines
* LayoutTests/fast/table/table-different-overflow-values.html: Add Test Case
* LayoutTests/fast/table/table-different-overflow-values-expected.txt: Add Test Case Expectation
* LayoutTests/fast/table/table-different-overflow-values-2.html: Add Test Case
* LayoutTests/fast/table/table-different-overflow-values-2-expected.txt: Add Test Case Expectation
* LayoutTests/fast/table/overflowScroll.html: Add Test Case
* LayoutTests/fast/table/overflowScroll-expected.html: Add Test Case Expectation
* LayoutTests/fast/table/overflowScroll-display-block.html: Add Test Case
* LayoutTests/fast/table/overflowScroll-display-block-expected.html: Add Test Case Expectation