Add String / StringView find() overloads that take in a LChar / char instead of a UChar https://bugs.webkit.org/show_bug.cgi?id=252442
Reviewed by Yusuke Suzuki.

Add String / StringView find() overloads that take in a LChar / char instead of
a UChar. It is pretty common to see code in WebKit doing `string.find(':')`.
When the input string is 8-bit, converting ':' to a UChar can be inefficient as
we have optimized versions of StringImpl::find() / WTF::find() that take in a
LChar/char.

Got the idea from:
https://chromium.googlesource.com/chromium/blink/+/c70fbd89e637c6fbd7bfeb17e6f5b2ff2dae6f16

* Source/WTF/wtf/text/StringView.h:
(WTF::StringView::find const):
* Source/WTF/wtf/text/WTFString.h:

Canonical link: https://commits.webkit.org/260415@main