Take legend element into account in propagateRepaintToParentWithOutlineAutoIfNeeded https://bugs.webkit.org/show_bug.cgi?id=251381 rdar://104813886
Reviewed by Alan Baradlay.

In change r259412 logic was introduced for spanner placeholders and a check was done
to see if the previous sibling renderer is a column set. However legends are kept out of
column flows and thus may also have a column set as previous sibling, in this case we
don't want to enter the spanner placeholder logic.

* LayoutTests/fast/multicol/legend-in-column-outline-auto-crash-expected.txt: Added.
* LayoutTests/fast/multicol/legend-in-column-outline-auto-crash.html: Added.
* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::RenderObject::propagateRepaintToParentWithOutlineAutoIfNeeded const):

Canonical link: https://commits.webkit.org/260286.11@webkit-2023.2-embargoed