Calculate auto margins on orthogonal tables correctly
https://bugs.webkit.org/show_bug.cgi?id=250206
rdar://problem/104223475

Reviewed by NOBODY (OOPS!).

This patch aligns WebKit to Blink / Chromium and Gecko / Firefox.

Partial Merge (only RenderTable.cpp) - https://chromium.googlesource.com/chromium/blink/+/977524d1fa2ea6e758b8e25a7a08473e1559f9b8

When a table is in an orthogonal flow to its containing block and it has
auto margins its important that we know the logical height of the table
if we are to get the margins right. So don't call 'updateLogicalHeight'
in RenderTable::layout() until we've computed the logical height of the
table. (For normal flow tables, updateLogicalHeight() is just a way of obtaining
and updating the margins.)

* Source/WebCore/rendering/RenderTable.cpp:
(RenderTable::layout): Just call "updateLogicalHeight" later without "if" condition
* LayoutTests/fast/css/vertical-lr-table-bfc-auto-margins-beside-float.html: Add Test Case
* LayoutTests/fast/css/vertical-lr-table-bfc-auto-margins-beside-float-expected.html: Add Test Case Expectation