[WebCore] Shrink FontVariantAlternates https://bugs.webkit.org/show_bug.cgi?id=252479 rdar://105594318
Reviewed by Tim Nguyen.

Reduce the size of `FontVariantAlternates` from 120 to 80 bytes by replacing all
the instance of `std::optional<String>` in `FontVariantAlternatesValues` with
`String`. `String` can already distinguish between an empty string and a null
string, so it seems redundant to wrap it in an optinal. Combined with #10722,
this further reduces the size of `StyleInheritedData` to 264 bytes.

* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
(WebCore::CSSPropertyParserHelpers::consumeFontVariantAlternates):
* Source/WebCore/platform/text/TextFlags.cpp:
(WebCore::operator<<):
* Source/WebCore/platform/text/TextFlags.h:

Canonical link: https://commits.webkit.org/260465@main