REGRESSION(257538@main): CSS opacity doesn't work on dialog::backdrop https://bugs.webkit.org/show_bug.cgi?id=252536 rdar://105636947
Reviewed by Simon Fraser.

The background layer functionality is to avoid flashes of white when rotating the screen on mobile device.
However, it breaks the opacity property, so only restrict it to the fullscreen element's backdrop, which is the only main
case that needs this polish (e.g. rotating while watching a fullscreen video).

* LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-backdrop-opacity-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-backdrop-opacity-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-backdrop-opacity.html: Added.
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::RenderLayerBacking):

Canonical link: https://commits.webkit.org/260556@main