[WebAuthn] Only pass along destinationSiteForCrossSiteAssertion if cross-site and normalize its format https://bugs.webkit.org/show_bug.cgi?id=252609 rdar://105694950
Reviewed by NOBODY (OOPS!).

Currently we pass a value for destinationSiteForCrossSiteAssertion in same-site, cross-origin cases. This
does not match the name of the variable. To fix this, we start passing it only when the value is cross-site.
We also normalize its value to match how to pass the RP ID.

* Source/WebKit/UIProcess/WebAuthentication/Cocoa/WebAuthenticatorCoordinatorProxy.mm:
(WebKit::configureAssertionOptions):
* Source/WebKit/WebProcess/WebAuthentication/WebAuthenticatorCoordinator.cpp:
(WebKit::WebAuthenticatorCoordinator::getAssertion):