Regression(252852@main) Games on kongregate.com are no longer loading https://bugs.webkit.org/show_bug.cgi?id=252636 rdar://104392542
Reviewed by Ryosuke Niwa.

Do not lazy load iframes that have no valid URL or "about:blank" as URL. Lazily
loading such iframes has no performance benefits and can actually cause breakage
as JS (like on kongregate.com) may expect them to load synchronously.

Note that we have a more general problem where we fail to create the Frame &
initial empty document when we decide to lazy load an iframe. This is not correct,
and we should only delay the navigation to the frame URL *after* the creation of
the initial empty document. However, fixing this is a larger change and this is
not a regression from 252852@main. As a result, I am merely adding a FIXME comment
in this patch and will address separately.

* LayoutTests/fast/frames/about-blank-frame-no-lazy-loading-expected.txt: Added.
* LayoutTests/fast/frames/about-blank-frame-no-lazy-loading.html: Added.
Add layout test coverage.

* Source/WebCore/html/HTMLFrameElementBase.cpp:
(WebCore::HTMLFrameElementBase::openURL):
* Source/WebCore/html/HTMLIFrameElement.cpp:
(WebCore::isFrameLazyLoadable):
(WebCore::HTMLIFrameElement::shouldLoadFrameLazily):

Canonical link: https://commits.webkit.org/260612@main