webgl/webgl-metal-disabled.html and webgl/webgl-via-metal-flag-off.html trigger GPU process crash https://bugs.webkit.org/show_bug.cgi?id=252659 rdar://102072583
Reviewed by Antti Koivisto.

Disable the tests that test running Cocoa WebGL without Metal.
Enable the tests only on BigSur, as that is the only platform that
ships devices that use WebGL with ANGLE/OpenGL.

The reason for the crash is:

    if (( "${TARGET_MAC_OS_X_VERSION_MAJOR}" >= 130000 ))
        then
            plistbuddy Add :com.apple.private.gpu-restricted bool YES
        fi

E.g. the new OSes have an entitlement disabling the GPU access needed
for the OpenGL and this is applied to all processes.

* LayoutTests/TestExpectations:
* LayoutTests/platform/mac-bigsur/TestExpectations: Added.

Canonical link: https://commits.webkit.org/260668@main