[WPE] built-product-archive should not unconditionally include the Tools and Source directories https://bugs.webkit.org/show_bug.cgi?id=252711
Reviewed by Carlos Garcia Campos.

The built products generated for WPE were including some stamp files
from the Cog build that they shouldn't. Like all the ones inside the
sub-directory `Tools/cog-prefix/src/cog-stamp`

This was causing an error when downloading a built-product and then
trying to do a build without cleaning the downloaded built-product
(basically what bots do). The error is caused because CMake was
tricked to think that it has already executed the steps to download
Cog (because the stamp files were present) when it still hasn't.

This stamp files were included in the first place by mistake when
trying to include the debug fission files (`.dwo`), because the
current code was adding the main top-level directories of `Tools`
and `Source` without further considerations (adding everything
inside this directories instead of just the `.dwo` files)

This patch changes the code to explicitly add each individual
`.dwo` file instead of adding the whole directories.

Since this list can be quite large, we pass it to the zip
command via stdin, to avoid issues with the ARG_MAX limit.

* Tools/CISupport/built-product-archive:

Canonical link: https://commits.webkit.org/260672@main