NotificationEvent interface exists but Notification interface does not on iOS 16.4 https://bugs.webkit.org/show_bug.cgi?id=252625 rdar://105697554
Reviewed by Chris Dumez.

On iOS, NotificationEvent is enabled by default but Notification is disabled by default. This
doesn't make any sense, since a NotificationEvent contains a Notification. Furthermore, you can't
actually even get a NotificationEvent on iOS with a default browser since the supporting machinery
to generate and fire the notification event only exists in Web.app. So it should definitely not
default to true.

To fix this, we only enable NotificationEvent where we used to (i.e. on macOS >= 13). Web.app will
enable this preference at runtime so that Web Push and NotificationEvent work as expected in that
app.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebKit/UIProcess/API/Cocoa/WKPreferences.mm:
(-[WKPreferences _setNotificationEventEnabled:]):
(-[WKPreferences _notificationEventEnabled]):
* Source/WebKit/UIProcess/API/Cocoa/WKPreferencesPrivate.h:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/PushAPI.mm:

Canonical link: https://commits.webkit.org/260685@main