Allow use-after-move to verify nullness in TestWTF.WTF_CheckedPtr.CheckedRef API test https://bugs.webkit.org/show_bug.cgi?id=252771
Reviewed by Ryosuke Niwa.

When you move from a CheckedRef, we call __asan_poison_memory_region to indicate you shouldn't use it,
and you shouldn't.  This test verifies that if you do call CheckedRef::ptr() you get null, which is
fine for the test to do.  We don't want to unpoison the memory in the production code, so I just call
__asan_unpoison_memory_region to allow the test to verify the use-after-move value.

* Tools/TestWebKitAPI/Tests/WTF/CheckedPtr.cpp:
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/260711@main