Some SystemPreview actions are not being detected https://bugs.webkit.org/show_bug.cgi?id=252780 rdar://105800886
Reviewed by Darin Adler.

We're getting reports of some SystemPreviews not detecting the user action.
We've been unable to reproduce, so this is adding some release logging to help
us narrow down the problem.

* Source/WebKit/Platform/Logging.h:
* Source/WebKit/UIProcess/Cocoa/SystemPreviewControllerCocoa.mm: Add logging when a user
taps on a link, it finishes loading, the user clicks, and the preview ends.

Canonical link: https://commits.webkit.org/260727@main