Clarify WebPageProxy wheel event code for UI-side compositing https://bugs.webkit.org/show_bug.cgi?id=252703 rdar://105752960
Reviewed by Tim Horton.

Make use of DrawingAreaProxy::shouldSendWheelEventsToEventDispatcher() to conditionalize
code in WebPageProxy, making it clear that we only call into m_scrollingCoordinatorProxy if we're
not sending events to the EventDispatcher, and that we don't need to send ScrollingAccelerationCurves
to the web process with UI-side compositing.

Also dispatchWheelEventWithoutScrolling() is only used on iOS, so #ifdef it for clarity.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::handleWheelEvent):
(WebKit::WebPageProxy::sendWheelEvent):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:

Canonical link: https://commits.webkit.org/260752@main